Refactored eval function
The eval function was renamed to eval_sign to take in account for
the evaluation with additional sign taking.
The new function eval now evaluates a set of challenges with a
weight array and does NOT take the sign of the responses.
This is particularly useful for the learning with LR.